requirements.txt
configparser==3.5.0
Pillow==5.2.0
pytz==2018.5
